(Issue #330)
This is the foundation for the exception class, which is set to handle
errors and relay specific messages about what went wrong to the user. If
this is approved, our next steps will be making the messages in the
catch blocks our own, while still preserving the stack track for the
developers.